+++ src/components/Button.tsx
@@ -4,7 +4,11 @@ import StyledButton from '@components/styles/StyledButton';
 interface ButtonProps { kind: string }
 
 export class Button extends React.Component <ButtonProps, {}> {
+  public static defaultProps = {
+    kind: "start",
+  };
+
   public render(): JSX.Element {
-    return <button>{ this.props.kind }</button>;
+    return <StyledButton>{ this.props.kind }</StyledButton>;
   }
 }